moved KeyHandler: utility -> utility/singleton
