+++ src/main/java/playerquests/Core.java
@@ -4,7 +4,7 @@ import org.bukkit.plugin.Plugin; // export the plugin for use elsewhere
 import org.bukkit.plugin.java.JavaPlugin; // essential for initialising the plugin
 
 import playerquests.client.chat.command.Commandplayerquest; // command to enter the main GUI
-import playerquests.utility.KeyHandler; // special class for using keys to reference any method
+import playerquests.utility.singleton.KeyHandler; // special class for using keys to reference any method
 
 /**
  * Entry point for the plugin.
